+++ src/commands/compile.ts
@@ -83,7 +83,7 @@ export default function compile(
       })
       .then(finished)
       .catch(onError)
-      .finally(updateSaveHistory);
+      .then(updateSaveHistory);
   } else if (folderToolingType && toolingType === undefined) {
     // This process uses the Metadata API to deploy specific files
     // This is where we extend it to create any kind of metadata
@@ -94,7 +94,7 @@ export default function compile(
       .then(reportMetadataResults)
       .then(finished)
       .catch(onError)
-      .finally(updateSaveHistory);
+      .then(updateSaveHistory);
   } else if (toolingType === undefined) {
     return Promise.reject({ message: 'Metadata Describe Error. Please try again.' });
   } else if (toolingType === 'AuraDefinition') {
@@ -102,12 +102,12 @@ export default function compile(
     return saveAura(document, toolingType, Metadata, forceCompile)
       .then(finished)
       .catch(onError)
-      .finally(updateSaveHistory);
+      .then(updateSaveHistory);
   } else if (toolingType === 'LightningComponentResource') {
     return saveLWC(document, toolingType, forceCompile)
       .then(finished)
       .catch(onError)
-      .finally(updateSaveHistory);
+      .then(updateSaveHistory);
   } else {
     // This process uses the Tooling API to compile special files like Classes, Triggers, Pages, and Components
     return saveApex(document, toolingType, Metadata, forceCompile)
@@ -118,7 +118,7 @@ export default function compile(
         });
       })
       .catch(onError)
-      .finally(updateSaveHistory);
+      .then(updateSaveHistory);
   }
 
   // =======================================================================================================================================
@@ -243,6 +243,16 @@ export default function compile(
     } else if (res.State === 'Error') {
       onError(res);
       failures++;
+    } else if (res.status === 'Failed') {
+      if (res.message) {
+        errMessages.push(res.message);
+      } else {
+        // capture a failed deployment there is no message returned, so guide user to view in Salesforce
+        errMessages.push(
+          'Deployment failed. Please view the details in the deployment status section in Salesforce.'
+        );
+      }
+      return false; // don't show the failed build error
     }
 
     if (failures === 0) {
@@ -252,8 +262,8 @@ export default function compile(
           res.records[0].DeployDetails.componentSuccesses[0].id
         );
         if (fcfile) {
-          var fcMem: forceCode.IWorkspaceMember | undefined = fcfile.getWsMember();
-          if (fcMem) {
+          var fcMem: forceCode.IWorkspaceMember = fcfile.getWsMember();
+          if (fcMem.coverage) {
             fcMem.coverage = undefined;
             fcfile.updateWsMember(fcMem);
           }
@@ -318,7 +328,7 @@ export default function compile(
 
   function updateSaveHistory(): boolean {
     saveHistoryService.addSaveResult({
-      fileName: parsers.getWholeFileName(document) || 'UKNOWN',
+      fileName: parsers.getWholeFileName(document) || 'UNKNOWN',
       path: document.fileName,
       success: errMessages.length === 0,
       messages: errMessages,