+++ src/commands/saveLWC.ts
@@ -1,34 +1,29 @@
 import * as vscode from 'vscode';
 import * as parsers from './../parsers';
 import { saveService } from '../services';
-import diff from './diff';
 import * as path from 'path';
 import { createPackageXML, deployFiles } from './deploy';
 
 // =======================================================================================================================================
-// ================================                Lightning Components               ===========================================
+// ================================                Lightning Web Components               ================================================
 // =======================================================================================================================================
 export function saveLWC(
   document: vscode.TextDocument,
   toolingType: string,
   forceCompile?: boolean
 ): Promise<any> {
-  const fileName: string | undefined = document.fileName.split(path.sep).pop();
   const name: string | undefined = parsers.getName(document, toolingType);
-  const Format: string | undefined = parsers.getFileExtension(document);
-  var Source: string = document.getText();
-  var currentObjectDefinition: any = undefined;
 
   return Promise.resolve(vscode.window.forceCode)
     .then(getLWCBundle)
     .then(ensureLWCBundle)
     .then(bundle => {
-      if (bundle !== true) {
+      if (Array.isArray(bundle)) {
         return getLWCDefinition(bundle).then((definitions: any) =>
           upsertLWCDefinition(definitions, bundle)
         );
       } else {
-        return Promise.resolve();
+        return Promise.resolve(bundle);
       }
     });
 
@@ -42,15 +37,7 @@ export function saveLWC(
     // If the Bundle doesn't exist, create it, else Do nothing
     if (name && (results.length === 0 || !results[0])) {
       // Create LWC Bundle
-      return createPackageXML([document.fileName], vscode.window.forceCode.storageRoot).then(() => {
-        const files: string[] = [];
-        files.push(path.join('lwc', name));
-        files.push('package.xml');
-        return deployFiles(files, vscode.window.forceCode.storageRoot)
-          .then(() => {
-            return true;
-          });
-      });
+      return saveLWCPackage();
     } else {
       return results;
     }
@@ -62,61 +49,47 @@ export function saveLWC(
     });
   }
   function upsertLWCDefinition(definitions: any, bundle: any) {
-    // If the Definition doesn't exist, create it
-    var def: any[] = definitions.filter(
-      (result: any) => result.FilePath.split('/').pop() === fileName
-    );
-    currentObjectDefinition = def.length > 0 ? def[0] : undefined;
-    if (currentObjectDefinition !== undefined) {
-      const serverContents: string = currentObjectDefinition.Source;
-      if (!forceCompile && !saveService.compareContents(document, serverContents)) {
-        return vscode.window
-          .showWarningMessage('Someone has changed this file!', 'Diff', 'Overwrite')
-          .then(s => {
-            if (s === 'Diff') {
-              diff(document, true);
-              return {};
-            }
-            if (s === 'Overwrite') {
-              return updateLWC();
-            }
-            return {};
-          });
-      } else {
-        return updateLWC();
-      }
-    } else if (bundle[0]) {
-      var filePath = document.fileName.split(vscode.window.forceCode.projectRoot + path.sep).pop();
-      const errObj = {
-        State: 'Error',
-        message:
-          'Error: File not created on server either because the name of the file is incorrect or there are syntax errors.',
-      };
-      if (filePath) {
-        return vscode.window.forceCode.conn.tooling
-          .sobject('LightningComponentResource')
-          .create({
-            LightningComponentBundleId: bundle[0].Id ? bundle[0].Id : bundle[0].id,
-            Format,
-            Source,
-            FilePath: filePath.split('\\').join('/'),
-          })
-          .catch(err => {
-            return errObj;
-          });
-      } else {
-        return errObj;
-      }
+    const filePath: string | undefined = document.fileName.split(path.sep).shift();
+    if (!filePath) {
+      return undefined;
     }
-    return undefined;
+    // add files in the folder to PreSaveFiles so we can compare content
+    saveService.addFilesInFolder(filePath);
+    var changedFiles: any[] = definitions.filter(
+      def =>
+        !saveService.compareContents(path.join(filePath, def.FilePath.split('/').pop()), def.Source)
+    );
+
+    if (changedFiles.length !== 0) {
+      var changedFileNames: string = changedFiles
+        .map(file => file.FilePath.split('/').pop())
+        .join(',');
+      return vscode.window
+        .showWarningMessage(
+          `Someone has changed ${changedFileNames} in this bundle! Overwrite?`,
+          'Yes',
+          'No'
+        )
+        .then(s => {
+          if (s === 'Yes') {
+            return saveLWCPackage();
+          }
+          return {};
+        });
+    } else {
+      return saveLWCPackage();
+    }
   }
 
-  function updateLWC() {
-    return vscode.window.forceCode.conn.tooling
-      .sobject('LightningComponentResource')
-      .update({ Id: currentObjectDefinition.Id, Source })
-      .catch(err => {
-        return { State: 'Error', message: err.message ? err.message : err };
-      });
+  function saveLWCPackage() {
+    if (!name) {
+      return undefined;
+    }
+    return createPackageXML([document.fileName], vscode.window.forceCode.storageRoot).then(() => {
+      const files: string[] = [];
+      files.push(path.join('lwc', name));
+      files.push('package.xml');
+      return deployFiles(files, vscode.window.forceCode.storageRoot);
+    });
   }
 }